+++ src/actinia_core/core/common/app.py
@@ -106,6 +106,7 @@ a cloud storage for download.
 """
 
 from flask_httpauth import HTTPBasicAuth
+from flask_httpauth import HTTPTokenAuth
 from flask_cors import CORS
 from flask import Flask
 from flask_restful_swagger_2 import Api
@@ -125,6 +126,7 @@ flask_app = Flask(__name__)
 flask_app.url_map.strict_slashes = False
 CORS(flask_app)
 
+
 flask_api = Api(
     flask_app,
     prefix=URL_PREFIX,
@@ -136,10 +138,15 @@ flask_api = Api(
     consumes=["application/gml+xml", "application/json"],
 )
 
-# Set the security definition in an unconventional way
-flask_api._swagger_object["securityDefinitions"] = {
-    "basicAuth": {"type": "basic"}
-}
-flask_api._swagger_object["security"] = [{"basicAuth": []}]
+# authentication method
+token_verification = True  # TODO from config
+if token_verification is True:
+    auth = HTTPTokenAuth(scheme='Bearer')
+else:
+    # Set the security definition in an unconventional way
+    flask_api._swagger_object["securityDefinitions"] = {
+        "basicAuth": {"type": "basic"}
+    }
+    flask_api._swagger_object["security"] = [{"basicAuth": []}]
 
-auth = HTTPBasicAuth()
+    auth = HTTPBasicAuth()
